refactor: renamed file and removed unneeded exports
