+++ motions/jumpToLink.ts
@@ -1,5 +1,5 @@
 import { MotionFn } from "./utils/defineObsidianVimMotion";
-import { jumpToPattern } from "./utils/getNthInstanceOfPattern";
+import { jumpToPattern } from "./utils/jumpToPattern";
 
 const LINK_REGEX = /\[\[[^\]\]]+?\]\]/g;
